refactor: extract mobile navigation to separate SiteBottomNav component
Move mobile navigation from SiteHeader to dedicated SiteBottomNav
component to improve code organization and separation of concerns.
Update styling for better mobile experience and adjust scroll-to-top
positioning to accommodate bottom navigation.
