+++ src/app/(app)/layout.tsx
@@ -1,5 +1,6 @@
 import dynamic from "next/dynamic"
 
+import { SiteBottomNav } from "@/components/site-bottom-nav"
 import { SiteFooter } from "@/components/site-footer"
 import { SiteHeader } from "@/components/site-header"
 
@@ -13,6 +14,11 @@ export default function AppLayout({ children }: { children: React.ReactNode }) {
       <SiteHeader />
       <main className="max-w-screen overflow-x-clip px-2">{children}</main>
       <SiteFooter />
+      <div
+        className="pointer-events-none fixed inset-x-0 bottom-0 z-50 h-[calc(--spacing(24)+env(safe-area-inset-bottom,0))] bg-linear-to-b from-transparent from-[calc(env(safe-area-inset-bottom,0%))] to-background mask-linear-[to_top,var(--background)_25%,transparent] backdrop-blur-[1px]"
+        aria-hidden
+      />
+      <SiteBottomNav />
       <ScrollToTop />
     </div>
   )
