+++ src/components/site-footer-brand.tsx
@@ -30,7 +30,7 @@ export function SiteFooterInteractiveLogotype() {
   }
 
   return (
-    <div className="screen-line-bottom after:z-1 after:bg-foreground/10">
+    <div className="screen-line-bottom after:z-1 after:bg-foreground/15">
       <div
         className="overflow-hidden"
         onMouseMove={handleMouseMove}
